third_party/blink/web_tests/flag-specific/highdpi/overflow/overflow-position-004-expected.txt

This is a testharness.js-based test.
FAIL absolute position with rtl scrollbars assert_equals: left expected 15 but got 14.666666984558105
Harness: the test ran to completion.
